r6502asm/pyproject.toml
[project]
name = "r6502"
version = "0.1.0"
description = "Core app"
requires-python = ">= 3.10"
dependencies = [
  "ophis-asm",
]

[project.optional-dependencies]
dev = [
  "autopep8",
  "basedpyright",
  "boto3-stubs",
  "mypy",
  "pycodestyle == 2.10.0",
  "pytest",
  "pytest-cov",
]

[project.scripts]
r6502asm = "r6502asm.__main__:cli_main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["r6502asm"]

[tool.setuptools.dynamic]
version = {attr = "r6502asm.__version__"}

[tool.setuptools.package-data]
"r6502asm" = ["py.typed"]
